Refactor optimization modules
