+++ dhnx/network.py
@@ -16,7 +16,7 @@ import os
 import pandas as pd
 
 from .graph import thermal_network_to_nx_graph
-from .optimization import optimize_operation, setup_optimise_investment, \
+from .optimization.optimization_models import optimize_operation, setup_optimise_investment, \
     solve_optimisation_investment
 from .helpers import Dict
 from .input_output import CSVNetworkImporter, CSVNetworkExporter, load_component_attrs
