+++ dhnx/optimization/dhs_nodes.py
@@ -13,8 +13,8 @@ SPDX-License-Identifier: MIT
 
 
 import oemof.solph as solph
-import dhnx.optimization_oemof_heatpipe as oh
-import dhnx.optimization_add_components as ac
+import dhnx.optimization.oemof_heatpipe as oh
+import dhnx.optimization.add_components as ac
 
 
 def add_nodes_dhs(opti_network, gd, nodes, busd):
